Fix hearing sometimes taking precedence over line of sight for zombies. General improvements and refactoring.

+++ Assets/Scripts/Utilities/HashAudiblesJob.cs
@@ -6,11 +6,11 @@ using Unity.Mathematics;
 [BurstCompile]
 public partial struct HashAudiblesJob : IJobEntity
 {
-    public NativeParallelMultiHashMap<int, int3>.ParallelWriter ParallelWriter;
+    public NativeParallelMultiHashMap<uint, int3>.ParallelWriter ParallelWriter;
 
     public void Execute(in Audible audible)
     {
-        var hash = (int)math.hash(audible.GridPositionValue);
+        var hash = math.hash(audible.GridPositionValue);
         ParallelWriter.Add(hash, audible.Target);
     }
 }
